taxitogether/requirements.txt
Django==1.6
Markdown==2.3.1
PIL==1.1.7
argparse==1.2.1
django-filter==0.7
djangorestframework==2.3.9
wsgiref==0.1.2
